Using ehcache instead of DB to hold user handles.
Instead of requesting the DB for the handle and update the DB set with the new timestamp for each request, switch this patch to an ehcache implementation. This reduces external communication and should reduce response time at all. The session idle timeout is now to configure via ehcache configuraton!
A fall back mechanism is implemented. If a handle is not in the ehcache than is the DB the fall back case. This is still required because of the persistent handles. If we have a solution to load the persistent handles into the cache than we can remove the fall back.


git-svn-id: http://escidev5.fiz-karlsruhe.de:8888/prototype/eSciDocCore@44224 5c398c58-3d0a-0410-a38c-d0825f6816e0